Update to code to make it more clean
