+++ src/app/visit-button/visit-button.component.ts
@@ -6,8 +6,12 @@ import { Component } from '@angular/core';
   styleUrls: ['./visit-button.component.css'],
 })
 export class VisitButtonComponent {
+  // Method to open Calendly in a new browser window
   openCalendly() {
+    // Calendly URL for booking a visit
     const calendlyURL = 'https://calendly.com/sillygeese341/book-visit-time';
+    
+    // Opening the Calendly URL in a new browser window or tab
     window.open(calendlyURL, '_blank');
   }
 }
